Add kernel parameter to enable i2c-6 pinctrl mapping
By default the pinctrl platform driver for Merrifield will register a
pinctrl mapping for i2c-6 that sets the correct pinmux once i2c-6 is
probed. If this is not desired, the following can be added to the kernel
command line: platform_mrfld_pinctrl.pinctrl_i2c6=n

Since the SCU driver is probed after the I2C bus drivers (due to the
order of PCI IDs), setting the pinmux for i2c-6 will fail since the SCU
driver must be probed to successfully set the bufcfg. As a workaround,
the I2C bus driver (i2c-designware-pci) can be compiled as a LKM and
inserted after all other PCI devices have been probed.